TGP-1333: added MaintainProfileConnector and refactored service to use it
